requirements.txt
numpy>=1.19.4
torch>=1.7.0
scikit-image==0.17.2
torchvision==0.8.1
tqdm>=4.52
ninja==1.10.0.post2
opencv-python==4.4.0.46
dlib>=19.18.0
hydra-core==1.0.6
